requirements.txt
arrow==1.3.0
asgiref==3.8.1
blessed==1.21.0
Django==5.2.1
django-multiselectfield==0.1.13
djangorestframework==3.16.0
mysqlclient==2.2.7
python-dateutil==2.9.0.post0
redis==3.5.3
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
wcwidth==0.2.13
